+++ resources/js/app.js
@@ -239,6 +239,11 @@ function getTemp( data ) {
 			           data.StatusSNS.BMP280.Temperature + "°" + data.StatusSNS.TempUnit
 		           ) );
 	}
+	if ( data.StatusSNS.SI7021 !== undefined ) {
+		temp.push( (
+			           data.StatusSNS.SI7021.Temperature + "°" + data.StatusSNS.TempUnit
+		           ) );
+	}
 	
 	//console.log( temp );
 	
@@ -261,6 +266,11 @@ function getHumidity( data ) {
 	if ( data.StatusSNS.SHT3X !== undefined ) {
 		if ( data.StatusSNS.SHT3X.Humidity !== undefined ) {
 			humi.push( data.StatusSNS.SHT3X.Humidity + "%" );
+		}
+	}
+	if ( data.StatusSNS.SI7021 !== undefined ) {
+		if ( data.StatusSNS.SI7021.Humidity !== undefined ) {
+			humi.push( data.StatusSNS.SI7021.Humidity + "%" );
 		}
 	}